changes according to PR comments
